+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `VimwikiAll2HTML[!]`
+
+### Changed
+
+- Update documentation
+- Add `vimwiki.allhtml` (bool) to `scripts/pre-commit.sh`.
+
 ## [v1.0.2] - 2022-09-04
 
 ### Changed